Moved test case to new file
